+++ Tests/PayID/PayIDIntegrationTests.swift
@@ -2,23 +2,38 @@ import Foundation
 import XCTest
 import XpringKit
 
-extension PaymentPointer {
-  public static let testPointer = "alice$dev.payid.xpring.money"
+extension String {
+  public static let testPayID = "alice$dev.payid.xpring.money"
 
   /// A pay ID that does not exist.
-  public static let nonExistentPointer = "doesNotExist$dev.payid.xpring.money"
+  public static let nonExistentPayID = "doesNotExist$dev.payid.xpring.money"
 }
 
 /// Integration tests run against a live PayID service.
 final class PayIDIntegrationTests: XCTestCase {
-  func testResolvePaymentPointerKnownPointerMainnet() {
-    let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
-
+  func testResolvePayIDKnownPayIDMainnetSync() {
     // GIVEN a Pay ID that will resolve on Mainnet and a PayID client.
     let payIDClient = XRPPayIDClient(xrplNetwork: .main)
 
-    // WHEN it is resolved to an XRP address.
-    payIDClient.xrpAddress(for: .testPointer) { result in
+    // WHEN it is resolved to an XRP address synchronously.
+    let result = payIDClient.xrpAddress(for: .testPayID)
+
+    // THEN the address is the expected value.
+    switch result {
+    case .success(let resolvedAddress):
+      XCTAssertEqual(resolvedAddress, "X7zmKiqEhMznSXgj9cirEnD5sWo3iZSbeFRexSFN1xZ8Ktn")
+    case .failure(let error):
+      XCTFail("Failed to resolve address: \(error)")
+    }
+  }
+
+  func testResolvePayIDKnownPayIDMainnetAsync() {
+    // GIVEN a Pay ID that will resolve on Mainnet and a PayID client.
+    let payIDClient = XRPPayIDClient(xrplNetwork: .main)
+
+    // WHEN it is resolved to an XRP address asynchronously.
+    let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
+    payIDClient.xrpAddress(for: .testPayID) { result in
       // THEN the address is the expected value.
       switch result {
       case .success(let resolvedAddress):
@@ -26,7 +41,26 @@ final class PayIDIntegrationTests: XCTestCase {
       case .failure(let error):
         XCTFail("Failed to resolve address: \(error)")
       }
+      expectation.fulfill()
+    }
+    self.wait(for: [ expectation ], timeout: 10)
+  }
 
+  // TODO(keefertaylor): This should  be a unit test. Migrate when
+  // https://github.com/xpring-eng/XpringKit/pull/238 is landed.
+  func testResolveXRPAddressKnownPayIDMainnetOnCustomThread() {
+    let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
+
+    // GIVEN a PayID that will resolve on Mainnet, an XRPPayIDClient and a custom callback queue.
+    let queueLabel = "io.xpring.XpringKit.test"
+    let customCallbackQueue = DispatchQueue(label: queueLabel)
+    DispatchQueue.registerDetection(of: customCallbackQueue)
+    let payIDClient = XRPPayIDClient(xrplNetwork: .main)
+
+    // WHEN it is resolved to an address and provided a custom queue
+    // THEN the callback is performed on the custom queue.
+    payIDClient.xrpAddress(for: .testPayID, callbackQueue: customCallbackQueue) { _ in
+      XCTAssertEqual(DispatchQueue.currentQueueLabel, queueLabel)
       expectation.fulfill()
     }
 
@@ -35,17 +69,18 @@ final class PayIDIntegrationTests: XCTestCase {
 
   // TODO(keefertaylor): This should  be a unit test. Migrate when
   // https://github.com/xpring-eng/XpringKit/pull/238 is landed.
-  func testResolvePaymentPointerKnownPointerMainnetOnCustomThread() {
+  func testResolveAddressKnownPayIDMainnetOnCustomThread() {
     let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
 
-    // GIVEN a Pay ID that will resolve on Mainnet and a PayID client and a custom callback queue.
+    // GIVEN a PayID that will resolve on Mainnet, a PayIDClient and a custom callback queue.
     let queueLabel = "io.xpring.XpringKit.test"
     let customCallbackQueue = DispatchQueue(label: queueLabel)
     DispatchQueue.registerDetection(of: customCallbackQueue)
     let payIDClient = PayIDClient()
 
-    // WHEN it is resolved to an XRP address and provided a custom queue and not on the main thread.
-    payIDClient.cryptoAddress(for: .testPointer, on: "xrpl-main", callbackQueue: customCallbackQueue) { _ in
+    // WHEN it is resolved to an XRP address and provided a custom queue
+    // THEN the callback is performed on the custom queue.
+    payIDClient.cryptoAddress(for: .testPayID, on: "xrpl-main", callbackQueue: customCallbackQueue) { _ in
       XCTAssertEqual(DispatchQueue.currentQueueLabel, queueLabel)
       expectation.fulfill()
     }
@@ -53,14 +88,14 @@ final class PayIDIntegrationTests: XCTestCase {
     self.wait(for: [ expectation ], timeout: 10)
   }
 
-  func testResolvePaymentPointerKnownPointerTestnet() {
+  func testResolvePayIDKnownPayIDTestnet() {
     let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
 
     // GIVEN a Pay ID that will resolve on testnet and a PayID client.
     let payIDClient = XRPPayIDClient(xrplNetwork: .test)
 
     // WHEN it is resolved to an XRP address.
-    payIDClient.xrpAddress(for: .testPointer) { result in
+    payIDClient.xrpAddress(for: .testPayID) { result in
       // THEN the address is the expected value.
       switch result {
       case .success(let resolvedAddress):
@@ -75,14 +110,14 @@ final class PayIDIntegrationTests: XCTestCase {
     self.wait(for: [ expectation ], timeout: 10)
   }
 
-  func testResolvePaymentPointerKnownPointerDevnet() {
+  func testResolvePayIDKnownPayIDDevnet() {
     let expectation = XCTestExpectation(description: "resolveToXRP completion called.")
 
     // GIVEN a Pay ID that will not resolve on Devnet and a PayID client.
     let payIDClient = XRPPayIDClient(xrplNetwork: .dev)
 
     // WHEN it is resolved to an XRP address.
-    payIDClient.xrpAddress(for: .nonExistentPointer) { result in
+    payIDClient.xrpAddress(for: .nonExistentPayID) { result in
       // THEN the result contains an `mappingNotFound` error.
       switch result {
       case .success:
@@ -104,7 +139,7 @@ final class PayIDIntegrationTests: XCTestCase {
     // GIVEN a Pay ID that will resolve on BTC Testnet.
     // WHEN it is resolved to an XRP address
     let payIDClient = PayIDClient()
-    let result = payIDClient.cryptoAddress(for: .testPointer, on: "btc-testnet")
+    let result = payIDClient.cryptoAddress(for: .testPayID, on: "btc-testnet")
 
     // THEN the address is the expected value.
     switch result {
@@ -119,7 +154,7 @@ final class PayIDIntegrationTests: XCTestCase {
     // GIVEN a PayID with multiple addresses.
     // WHEN all addresses are synchronously resolved.
     let payIDClient = PayIDClient()
-    let result = payIDClient.allAddresses(for: .testPointer)
+    let result = payIDClient.allAddresses(for: .testPayID)
 
     // THEN multiple results are returned.
     switch result {
@@ -136,7 +171,7 @@ final class PayIDIntegrationTests: XCTestCase {
     // GIVEN a PayID with multiple addresses.
     // WHEN all addresses are synchronously resolved.
     let payIDClient = PayIDClient()
-    payIDClient.allAddresses(for: .testPointer) { result in
+    payIDClient.allAddresses(for: .testPayID) { result in
       // THEN multiple results are returned.
       switch result {
       case .success(let resolvedAddress):